requirements.txt
Flask
web3
tensorflow
python-dotenv 